+++ .gitignore
@@ -1,0 +1,2 @@
+auth filter.txt
+loginControler.txt
